requirements.txt
absl-py==1.3.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
connected-components-3d==3.10.2
contourpy==1.0.5
cycler==0.11.0
et-xmlfile==1.1.0
flatbuffers==22.10.26
fonttools==4.38.0
gast==0.4.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
idna==3.4
imageio==2.22.2
importlib-metadata==5.0.0
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
networkx==2.8.7
nibabel==4.0.2
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
path==16.5.0
Pillow==9.2.0
#pkg_resources==0.0.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
PyWavelets==1.4.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.3
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
threadpoolctl==3.1.0
tifffile==2022.10.10
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
XlsxWriter==3.0.3
zipp==3.10.0

czifile~=2019.7.2